Allow to work with longer strign (from u8 to u32)

+++ tfhe/examples/fhe_strings/server_key/contains.rs
@@ -140,7 +140,7 @@ mod tests {
                     let encrypted_s = CLIENT_KEY.encrypt_str(&$string_arg).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &$pattern_arg);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -149,7 +149,7 @@ mod tests {
                     let encrypted_s = CLIENT_KEY.encrypt_str_random_padding(&$string_arg, 2).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &$pattern_arg);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -159,7 +159,7 @@ mod tests {
 		    let encrypted_pattern = CLIENT_KEY.encrypt_ascii_char($pattern_arg as u8);
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &encrypted_pattern);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -169,7 +169,7 @@ mod tests {
 		    let encrypted_pattern = CLIENT_KEY.encrypt_ascii_char($pattern_arg as u8);
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &encrypted_pattern);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
             }
         };
@@ -185,7 +185,7 @@ mod tests {
                     let encrypted_s = CLIENT_KEY.encrypt_str(&$string_arg).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &$pattern_arg);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -194,7 +194,7 @@ mod tests {
                     let encrypted_s = CLIENT_KEY.encrypt_str_random_padding(&$string_arg, 2).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &$pattern_arg);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -204,7 +204,7 @@ mod tests {
 		    let encrypted_pattern = CLIENT_KEY.encrypt_str(&$pattern_arg).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &encrypted_pattern);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -214,7 +214,7 @@ mod tests {
 		    let encrypted_pattern = CLIENT_KEY.encrypt_str(&$pattern_arg).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &encrypted_pattern);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
 
 		#[test]
@@ -224,7 +224,7 @@ mod tests {
 		    let encrypted_pattern = CLIENT_KEY.encrypt_str_random_padding(&$pattern_arg, 2).unwrap();
                     let fhe_result = SERVER_KEY.$method(&encrypted_s, &encrypted_pattern);
                      let clear_fhe_result = CLIENT_KEY.decrypt_u8(&SERVER_KEY.bool_to_radix(&fhe_result));
-		    assert_eq!(std_result as u8, clear_fhe_result);
+		    assert_eq!(std_result as u32, clear_fhe_result);
 		}
             }
         };
